move straggling packages
