--- pkg/cmd/util/net.go
@@ -1,36 +1,0 @@
-package util
-
-import (
-	"strings"
-
-	"k8s.io/apimachinery/pkg/util/sets"
-)
-
-// HostnameMatchSpecCandidates returns a list of match specs that would match the provided hostname
-// Returns nil if len(hostname) == 0
-func HostnameMatchSpecCandidates(hostname string) []string {
-	if len(hostname) == 0 {
-		return nil
-	}
-
-	// Exact match has priority
-	candidates := []string{hostname}
-
-	// Replace successive labels in the name with wildcards, to require an exact match on number of
-	// path segments, because certificates cannot wildcard multiple levels of subdomains
-	//
-	// This is primarily to be consistent with tls.Config#getCertificate implementation
-	//
-	// It using a cert signed for *.foo.example.com and *.bar.example.com by specifying the name *.*.example.com
-	labels := strings.Split(hostname, ".")
-	for i := range labels {
-		labels[i] = "*"
-		candidates = append(candidates, strings.Join(labels, "."))
-	}
-	return candidates
-}
-
-// HostnameMatches returns true if the given hostname is matched by the given matchSpec
-func HostnameMatches(hostname string, matchSpec string) bool {
-	return sets.NewString(HostnameMatchSpecCandidates(hostname)...).Has(matchSpec)
-}
