+++ staging/src/github.com/openshift/openshift-apiserver/pkg/cmd/openshift-apiserver/openshiftapiserver/config.go
@@ -28,10 +28,10 @@ import (
 	"github.com/openshift/library-go/pkg/apiserver/apiserverconfig"
 	"github.com/openshift/library-go/pkg/config/helpers"
 	"github.com/openshift/library-go/pkg/config/serving"
+	"github.com/openshift/openshift-apiserver/pkg/cmd/openshift-apiserver/openshiftadmission"
+	"github.com/openshift/openshift-apiserver/pkg/cmd/openshift-apiserver/openshiftapiserver/configprocessing"
 	"github.com/openshift/openshift-apiserver/pkg/image/apiserver/registryhostname"
 	"github.com/openshift/openshift-apiserver/pkg/version"
-	"github.com/openshift/origin/pkg/cmd/openshift-apiserver/openshiftadmission"
-	"github.com/openshift/origin/pkg/cmd/openshift-apiserver/openshiftapiserver/configprocessing"
 	"k8s.io/kubernetes/openshift-kube-apiserver/configflags"
 )
