+++ test/util/server/deprecated_openshift/iputil/ip.go
@@ -1,4 +1,4 @@
-package util
+package iputil
 
 import (
 	"errors"
